# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategoryUnit = typeof CategoryUnit[keyof typeof CategoryUnit];


export const CategoryUnit = {
  kilo: 'kilo',
  box: 'box',
  piece: 'piece',
} as const;

export interface Category {
  id: number;
  name: string;
  unit: CategoryUnit;
  pricePerUnit: number;
  stockQuantity: number;
  lowStockThreshold?: number;
  isActive?: boolean;
}

export type CategoryInputUnit = typeof CategoryInputUnit[keyof typeof CategoryInputUnit];


export const CategoryInputUnit = {
  kilo: 'kilo',
  box: 'box',
  piece: 'piece',
} as const;

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  unit: CategoryInputUnit;
  /** @minimum 0 */
  pricePerUnit: number;
  /** @minimum 0 */
  stockQuantity: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
}

export type CategoryUpdateUnit = typeof CategoryUpdateUnit[keyof typeof CategoryUpdateUnit];


export const CategoryUpdateUnit = {
  kilo: 'kilo',
  box: 'box',
  piece: 'piece',
} as const;

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  unit?: CategoryUpdateUnit;
  /** @minimum 0 */
  pricePerUnit?: number;
  /** @minimum 0 */
  stockQuantity?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  isActive?: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
} as const;

export type OrderItemUnit = typeof OrderItemUnit[keyof typeof OrderItemUnit];


export const OrderItemUnit = {
  kilo: 'kilo',
  box: 'box',
  piece: 'piece',
} as const;

export interface OrderItem {
  id: number;
  categoryId: number;
  categoryName: string;
  unit: OrderItemUnit;
  quantity: number;
  subtotal: number;
}

export type OrderCreatedBy = typeof OrderCreatedBy[keyof typeof OrderCreatedBy];


export const OrderCreatedBy = {
  guest: 'guest',
  admin: 'admin',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  phoneNumber: string;
  pickupDate: string;
  pickupTime: string;
  status: OrderStatus;
  /** @nullable */
  notes: string | null;
  totalPrice: number;
  createdBy: OrderCreatedBy;
  createdAt: string;
  items: OrderItem[];
}

export type OrderInputCreatedBy = typeof OrderInputCreatedBy[keyof typeof OrderInputCreatedBy];


export const OrderInputCreatedBy = {
  guest: 'guest',
  admin: 'admin',
} as const;

export interface OrderItemInput {
  categoryId: number;
  /** @minimum 0.25 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 7 */
  phoneNumber: string;
  pickupDate: string;
  /** @minLength 1 */
  pickupTime: string;
  notes?: string;
  createdBy?: OrderInputCreatedBy;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderUpdate {
  /** @minLength 2 */
  customerName?: string;
  /** @minLength 7 */
  phoneNumber?: string;
  pickupDate?: string;
  pickupTime?: string;
  notes?: string;
  status?: OrderStatus;
}

export interface DashboardSummary {
  pendingOrders: number;
  todayOrders: number;
  acceptedRevenue: number;
  lowStockCount: number;
  totalOrders: number;
  /** @nullable */
  topCategory?: string | null;
  daysUntilEid?: number;
}

export interface DailyOrderPoint {
  date: string;
  orders: number;
  revenue: number;
}

export interface CategoryTotal {
  categoryName: string;
  quantity: number;
  revenue: number;
}

export interface StatusTotal {
  status: OrderStatus;
  count: number;
}

export interface DashboardAnalytics {
  dailyOrders: DailyOrderPoint[];
  categoryTotals: CategoryTotal[];
  statusTotals: StatusTotal[];
}

export type ListOrdersParams = {
status?: OrderStatus;
date?: string;
search?: string;
};

export type TrackOrderParams = {
orderNumber?: string;
phone?: string;
};

export type ExportOrdersParams = {
range?: ExportOrdersRange;
};

export type ExportOrdersRange = typeof ExportOrdersRange[keyof typeof ExportOrdersRange];


export const ExportOrdersRange = {
  day: 'day',
  week: 'week',
} as const;