# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryInputUnit } from './categoryInputUnit';

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  unit: CategoryInputUnit;
  /** @minimum 0 */
  pricePerUnit: number;
  /** @minimum 0 */
  stockQuantity: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
}