# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryUpdateUnit } from './categoryUpdateUnit';

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  unit?: CategoryUpdateUnit;
  /** @minimum 0 */
  pricePerUnit?: number;
  /** @minimum 0 */
  stockQuantity?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  isActive?: boolean;
}