# lib/api-zod/src/generated/types/dashboardAnalytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryTotal } from './categoryTotal';
import type { DailyOrderPoint } from './dailyOrderPoint';
import type { StatusTotal } from './statusTotal';

export interface DashboardAnalytics {
  dailyOrders: DailyOrderPoint[];
  categoryTotals: CategoryTotal[];
  statusTotals: StatusTotal[];
}