# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCreatedBy } from './orderCreatedBy';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  phoneNumber: string;
  pickupDate: Date;
  pickupTime: string;
  status: OrderStatus;
  /** @nullable */
  notes: string | null;
  totalPrice: number;
  createdBy: OrderCreatedBy;
  createdAt: Date;
  items: OrderItem[];
}