# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputCreatedBy } from './orderInputCreatedBy';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 7 */
  phoneNumber: string;
  pickupDate: Date;
  /** @minLength 1 */
  pickupTime: string;
  notes?: string;
  createdBy?: OrderInputCreatedBy;
  /** @minItems 1 */
  items: OrderItemInput[];
}