# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemUnit } from './orderItemUnit';

export interface OrderItem {
  id: number;
  categoryId: number;
  categoryName: string;
  unit: OrderItemUnit;
  quantity: number;
  subtotal: number;
}