# lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
} as const;